move function for sorting blocks by order and nesting level to ui library
